Move setup state classes out
