+++ MvvmCross/Core/MvxSetupStateEventArgs.cs
@@ -1,0 +1,20 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MS-PL license.
+// See the LICENSE file in the project root for more information.
+
+using System;
+
+namespace MvvmCross.Core
+{
+#nullable enable
+    public class MvxSetupStateEventArgs : EventArgs
+    {
+        public MvxSetupStateEventArgs(MvxSetupState setupState)
+        {
+            SetupState = setupState;
+        }
+
+        public MvxSetupState SetupState { get; }
+    }
+#nullable restore
+}
